refactor(model): split pager helpers (model_pager.go), tree stream glue (model_tree_stream.go), resource loading (model_resources.go), init (model_init.go), tables+new (model_tables.go, model_new.go); model.go trimmed; no behavior change
